+++ compiler/codegen_qbe/codegen.zig
@@ -398,6 +398,7 @@ pub const Codegen = struct {
 
     pub const genCall = calls.genCall;
     pub const genParseOrRaise = calls.genParseOrRaise;
+    pub const emitHpyErrorCheckOrRaise = calls.emitHpyErrorCheckOrRaise;
     pub const releaseTemporaryArgs = calls.releaseTemporaryArgs;
     pub const releaseIfTemporary = calls.releaseIfTemporary;
     pub const currentArena = calls.currentArena;